+++ index.ts
@@ -1,3 +1,4 @@
 import { Docker } from './nodes/Docker/Docker.node';
+import { DockerTrigger } from './nodes/DockerTrigger/DockerTrigger.node';
 
-export const nodeTypes = [Docker];
+export const nodeTypes = [Docker, DockerTrigger];